+++ database/migrations/2024_10_29_012141_create_retail_categories_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,12 +12,14 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('retail_categories', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('client_identifier')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('retail_categories')) {
+            Schema::create('retail_categories', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->string('client_identifier')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -24,6 +27,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('retail_categories');
+        if (Schema::hasTable('retail_categories')) {
+            $count = DB::table('retail_categories')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('retail_categories');
+            }
+        }
     }
 };
