+++ database/migrations/2024_10_29_012142_create_retail_items_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,19 +12,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('retail_items', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('sku');
-            $table->json('images')->nullable();
-            $table->integer('quantity');
-            $table->string('unit')->nullable();
-            $table->decimal('price');
-            $table->foreignId('category_id')->constrained('retail_categories');
-            $table->text('barcode')->nullable();
-            $table->string('client_identifier')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('retail_items')) {
+            Schema::create('retail_items', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->string('sku');
+                $table->json('images')->nullable();
+                $table->integer('quantity');
+                $table->string('unit')->nullable();
+                $table->decimal('price');
+                $table->foreignId('category_id')->constrained('retail_categories');
+                $table->text('barcode')->nullable();
+                $table->string('client_identifier')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -31,6 +34,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('retail_items');
+        if (Schema::hasTable('retail_items')) {
+            $count = DB::table('retail_items')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('retail_items');
+            }
+        }
     }
 };
