+++ database/migrations/2024_10_29_012249_create_fnb_menu_items_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,16 +12,19 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('fnb_menu_items', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->decimal('price', 8, 2);
-            $table->string('category');
-            $table->string('image')->nullable();
-            $table->string('public_image_url')->nullable();
-            $table->string('client_identifier')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('fnb_menu_items')) {
+            Schema::create('fnb_menu_items', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->decimal('price', 8, 2);
+                $table->string('category');
+                $table->string('image')->nullable();
+                $table->boolean('is_available_personal')->default(true);
+                $table->boolean('is_available_online')->default(true);
+                $table->string('client_identifier')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -28,6 +32,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('fnb_menu_items');
+        if (Schema::hasTable('fnb_menu_items')) {
+            $count = DB::table('fnb_menu_items')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('fnb_menu_items');
+            }
+        }
     }
 };
