+++ database/migrations/2024_10_29_012337_create_fnb_kitchen_orders_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,14 +12,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('fnb_kitchen_orders', function (Blueprint $table) {
-            $table->id();
-            $table->string('table_number');
-            $table->json('items');
-            $table->string('status');
-            $table->string('client_identifier');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('fnb_kitchen_orders')) {
+            Schema::create('fnb_kitchen_orders', function (Blueprint $table) {
+                $table->id();
+                $table->string('table_number');
+                $table->json('items');
+                $table->string('status');
+                $table->string('client_identifier');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -26,6 +29,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('fnb_kitchen_orders');
+        if (Schema::hasTable('fnb_kitchen_orders')) {
+            $count = DB::table('fnb_kitchen_orders')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('fnb_kitchen_orders');
+            }
+        }
     }
 };
