+++ database/migrations/2024_10_29_012352_create_fnb_sales_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,17 +12,19 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('fnb_sales', function (Blueprint $table) {
-            $table->id();
-            $table->string('table_number');
-            $table->string('items');
-            $table->decimal('subtotal', 10, 2);
-            $table->decimal('discount', 10, 2)->nullable();
-            $table->string('discount_type')->nullable();
-            $table->decimal('total', 10, 2);
-            $table->string('client_identifier')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('fnb_sales')) {
+            Schema::create('fnb_sales', function (Blueprint $table) {
+                $table->id();
+                $table->string('table_number');
+                $table->string('items');
+                $table->decimal('subtotal', 10, 2);
+                $table->decimal('discount', 10, 2)->nullable();
+                $table->string('discount_type')->nullable();
+                $table->decimal('total', 10, 2);
+                $table->string('client_identifier')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -29,6 +32,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('fnb_sales');
+        if (Schema::hasTable('fnb_sales')) {
+            $count = DB::table('fnb_sales')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('fnb_sales');
+            }
+        }
     }
 };
