+++ database/migrations/2024_10_29_012416_create_fnb_tables_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,15 +12,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('fnb_tables', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->integer('seats');
-            $table->enum('status', ['available', 'occupied', 'reserved', 'joined']);
-            $table->string('joined_with')->nullable();
-            $table->string('client_identifier')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('fnb_tables')) {
+            Schema::create('fnb_tables', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->integer('seats');
+                $table->enum('status', ['available', 'occupied', 'reserved', 'joined']);
+                $table->string('joined_with')->nullable();
+                $table->string('client_identifier')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -27,6 +30,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('fnb_tables');
+        if (Schema::hasTable('fnb_tables')) {
+            $count = DB::table('fnb_tables')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('fnb_tables');
+            }
+        }
     }
 };
