+++ database/migrations/2024_10_29_012433_create_fnb_reservations_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,19 +12,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('fnb_reservations', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('date');
-            $table->string('time');
-            $table->string('guests');
-            $table->string('table_id');
-            $table->string('notes')->nullable();
-            $table->string('contact')->nullable();
-            $table->string('status');
-            $table->string('client_identifier');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('fnb_reservations')) {
+            Schema::create('fnb_reservations', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->string('date');
+                $table->string('time');
+                $table->string('guests');
+                $table->string('table_id');
+                $table->string('notes')->nullable();
+                $table->string('contact')->nullable();
+                $table->string('status');
+                $table->string('client_identifier');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -31,6 +34,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('fnb_reservations');
+        if (Schema::hasTable('fnb_reservations')) {
+            $count = DB::table('fnb_reservations')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('fnb_reservations');
+            }
+        }
     }
 };
