+++ database/migrations/2024_10_29_012634_create_transportation_fleets_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,10 +12,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('transportation_fleets', function (Blueprint $table) {
-            $table->id();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('transportation_fleets')) {
+            Schema::create('transportation_fleets', function (Blueprint $table) {
+                $table->id();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -22,6 +25,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('transportation_fleets');
+        if (Schema::hasTable('transportation_fleets')) {
+            $count = DB::table('transportation_fleets')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('transportation_fleets');
+            }
+        }
     }
 };
