+++ database/migrations/2024_10_29_012857_create_patient_bills_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,16 +12,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('patient_bills', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('patient_id')->constrained('patients')->onDelete('cascade');
-            $table->string('bill_number')->nullable();
-            $table->string('bill_date');
-            $table->decimal('bill_amount', 10, 2);
-            $table->string('bill_status')->nullable();
-            $table->string('bill_details')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('patient_bills')) {
+            Schema::create('patient_bills', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('patient_id')->constrained('patients')->onDelete('cascade');
+                $table->string('bill_number')->nullable();
+                $table->string('bill_date');
+                $table->decimal('bill_amount', 10, 2);
+                $table->string('bill_status')->nullable();
+                $table->string('bill_details')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -28,6 +31,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('patient_bills');
+        if (Schema::hasTable('patient_bills')) {
+            $count = DB::table('patient_bills')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('patient_bills');
+            }
+        }
     }
 };
