+++ database/migrations/2024_10_29_012905_create_patient_checkups_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
 return new class extends Migration
 {
@@ -11,15 +12,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('patient_checkups', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('patient_id')->constrained('patients')->onDelete('cascade');
-            $table->string('checkup_date');
-            $table->string('diagnosis')->nullable();
-            $table->string('treatment')->nullable();
-            $table->string('notes')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('patient_checkups')) {
+            Schema::create('patient_checkups', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('patient_id')->constrained('patients')->onDelete('cascade');
+                $table->string('checkup_date');
+                $table->string('diagnosis')->nullable();
+                $table->string('treatment')->nullable();
+                $table->string('notes')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -27,6 +30,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('patient_checkups');
+        if (Schema::hasTable('patient_checkups')) {
+            $count = DB::table('patient_checkups')->count();
+            if ($count === 0) {
+                Schema::dropIfExists('patient_checkups');
+            }
+        }
     }
 };
